requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
boto3==1.40.8
botocore==1.40.8
certifi==2025.8.3
charset-normalizer==3.4.3
diff-match-patch==20241021
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
jmespath==1.0.1
mysqlclient==2.2.7
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
gunicorn==21.2.0